This commit makes test_betweenness_centrality pass.

There was an issue with type dwimming.

The expected result has a specific type.

int is more specific than float.

We didn't specify that previously.

This made the type for the generated PythonNodes instance be the more general type that is float by default.

This type difference was causing the test failure.

+++ metagraph/tests/algorithms/test_betweenness_centrality.py
@@ -42,7 +42,9 @@ v        v /        v
         7: 0.0,
         6: 11.0,
     }
-    expected_answer = dpr.wrapper.Nodes.PythonNodes(expected_answer_unwrapped)
+    expected_answer = dpr.wrapper.Nodes.PythonNodes(
+        expected_answer_unwrapped, dtype="int"
+    )
     MultiVerify(
         dpr,
         "vertex_ranking.betweenness_centrality",
